Removed all the initEnv stuff; added option to extend the classpath.
The initEnv code was confusing-- I believe it is so you can source gsh.sh to
get a proper environment.  If so, you can still do that.  It doesn't make
sense for the script engine wrappers.

The -c option now lets you extend the CLASSPATH without modifying your local
environment.